+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v0.2.0 (Not released)
 
-* Add methods: `getAutoZoom`, `setAutoZoom`
+* Add methods: `getAutoZoom`, `setAutoZoom`, `loadMusicXML`, `getMusicXML`
 * Add events: `scoreLoaded`, `cursorPosition`, `rangeSelection`, `pause`, `stop`, `fullscreen`, `print`
 
 ## v0.1.0