+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v0.2.0 (Not released)
+## v0.2.0
 
 * Add methods: `getAutoZoom`, `setAutoZoom`, `loadMusicXML`, `getMusicXML`
 * Add events: `scoreLoaded`, `cursorPosition`, `rangeSelection`, `pause`, `stop`, `fullscreen`, `print`